+++ .trebired/logger/config.ts
@@ -1,4 +1,4 @@
 export default {
-  forVersion: "2.5.26",
+  forVersion: "2.5.28",
   prefix: "trebired.startup",
 };